refactor: move functions related with pr
